Make the release survive its own rolling replacement
A deploy readiness pass found three things that would have broken
production, and all three are the same shape: a migration that is
correct for the code shipping with it and wrong for the code already
running beside it.

**The pairing column.** Dropping `machine_pairings.user_id` is
correct — nothing writes or reads it, and the owner is reachable
through the computer. But migrations run on the first replaced node
while the other two still serve the previous release, and that release
declares `belongs_to :user`, which puts the column into every
generated SELECT. The drop would have broken every read and write of
that table on the un-replaced nodes: CLI device pairing down for two
thirds of traffic, then one third, for the length of the roll. The
drop is deferred to the release after the one that stopped writing it.

**The terminal-thread constraint.** Requiring `report_type` on a
terminal thread is what THREAD-003 promises, and the previous release
sets only status, report, report_digest, and completed_at when a
thread finishes. Requiring it here would stop old nodes closing,
cancelling, or expiring any thread mid-roll. The database admits a
missing report type for one release; the application supplies one on
every path it owns, and a later migration tightens the column.

**The silent narrowing.** `modify :max_total_tokens, :integer` on a
column created as `bigint` is not a nullability change — Ecto always
emits the type clause, so it issues ALTER COLUMN TYPE. That is a full
table rewrite under ACCESS EXCLUSIVE, it caps a grant at ~2.1 billion
tokens and ~$2,147 forever, the `down` does not restore `bigint`, and
it aborts outright on any row already above the limit — with
migrate-on-boot, that is a node that never starts serving. The type is
preserved in both directions; only nullability changes.

**And the credential that would have taken the default lane down.**
`vercel_gateway_api_key` was wired in the dev branch of runtime.exs
only, so in production the catalog's first entry — what a caller
naming no model gets, which is the published CLI's default path —
would have refused every call with `model_unavailable`. The prod
branch reads `AI_GATEWAY_API_KEY` now.

Verified against a database rebuilt from every migration in order, on
an isolated partition rather than the shared one: 4,535 tests green,
`inference_grants` ceilings still `bigint`, and the pairing column
still present.

+++ priv/repo/migrations/20260824231951_allow_unbounded_inference_grants.exs
@@ -21,9 +21,15 @@ defmodule OpenAgents.Repo.Migrations.AllowUnboundedInferenceGrants do
 
   def up do
     alter table(:inference_grants) do
-      modify :max_total_tokens, :integer, null: true
+      # The declared type must match what the table already holds. Ecto's
+      # Postgres adapter always emits the type clause on `modify`, so naming
+      # `:integer` here would issue `ALTER COLUMN TYPE integer` against columns
+      # created as `:bigint` — a full table rewrite that also caps a grant at
+      # ~2.1 billion tokens and ~$2,147 forever, and that aborts the migration
+      # outright on any row already above it. Only nullability is changing.
+      modify :max_total_tokens, :bigint, null: true
       modify :max_calls, :integer, null: true
-      modify :max_cost_microusd, :integer, null: true
+      modify :max_cost_microusd, :bigint, null: true
     end
 
     execute("""
@@ -52,9 +58,9 @@ defmodule OpenAgents.Repo.Migrations.AllowUnboundedInferenceGrants do
     """)
 
     alter table(:inference_grants) do
-      modify :max_total_tokens, :integer, null: false
+      modify :max_total_tokens, :bigint, null: false
       modify :max_calls, :integer, null: false
-      modify :max_cost_microusd, :integer, null: false
+      modify :max_cost_microusd, :bigint, null: false
     end
 
     execute(guard("<>"))
